// Front-End/src/app/model/Transaction/transaction.model.ts
export class Transaction {
    emitterId: number = 0;
    emitterFirstname: string = "";
    emitterLastname: string = "";
    receiverId: number = 0;
    receiverLastname: string = "";
    receiverFirstname: string = "";
    date: string = "";
    description: string = "";
    amount: number = 0;
}